{{cookiecutter.repo_name}}/requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7.1
django-discover-runner==1.0
Sphinx

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.0.1

# use a .env file like foreman
django-dotenv